+++ videoswin/__init__.py
@@ -1,1 +1,2 @@
+from .backbone import VideoSwinBackbone  # noqa: E501
 from .model import VideoSwinB, VideoSwinS, VideoSwinT  # noqa: E501
